修改显示 使用timetick_default 程序减少2K。 修改信号的显示

+++ bsp/stm32f40x_car/applications/jt808_param.h
@@ -117,7 +117,7 @@ typedef struct _jt808_param
 	char	id_0xF001[32];      /*0xF001 终端型号 20byte*/
 	char	id_0xF002[32];      /*0xF002 终端ID 7byte*/
 	char	id_0xF003[32];      /*0xF003 鉴权码*/
-	uint8_t id_0xF004;          /*0xF004 终端类型*/
+	uint16_t id_0xF004;          /*0xF004 终端类型*/
 	char	id_0xF005[32];      /*0xF005 车辆标识,VIN*/
 	char	id_0xF006[32];      /*0xF006 车辆标识,车牌号*/
 	uint8_t id_0xF007;          /*0xF007 车牌颜色*/
